Improve modbus
