+++ detect/ping.go
