+++ detect/registry.go
