+++ detect/sma.go
