+++ detect/tasklist.go
